Bug fixes and started making tutorial
Ragdoll bounce direction was originally based on mouse instead of facing direction and changed bounce effects to dropdown instead of checkmarks. Also started on tutorial level and so far added a backdrop, and a script where the words appear at a specified distance and can be used for other game objects in the future as well.